Reference commit in Node core that altered behaviour in v6
I bisected the Node.js core tree, and apparently
nodejs/node@b488b19eaf2b2e7a3ca5eccd2445e245847a5f76
is the commit that makes the bug occur more frequently in v6.

I re-tested that multiple times, and could produce the difference
rather reliably. This also makes sense because it’s a semver-major
change that was not included in any v5 version and was primarily
a performance change.